complete hangman object and moved to hangman_game
